# perf: Improve resmap.Append performance by using map lookup