@chlunde 2023-10-30
_A long time ago, I wrote_
> FYI, I took another look at the source code, and this approach does not work, as each resource might be modified without the resmap knowing about it.
> I think the resmap needs to be informed about all mutations to the objects to ensure the CurId is kept up to date. I'll propose some alternatives for this later when I have more time.

I will use another angle to attack this problem, and since It is a very different change I will create a new PR. I'll create the PR once a few of the other perf PRs are merged to avoid conflicts.